feat(standard): implement opt-in Standard JSON Schema coercion
