+++ packages/arkenv/src/coercion/shared.ts
@@ -1,0 +1,298 @@
+import { coerceBoolean, coerceDate, coerceJson, coerceNumber } from "./morphs.ts";
+
+/**
+ * A marker used in the coercion path to indicate that the target
+ * is the *elements* of an array, rather than the array property itself.
+ */
+export const ARRAY_ITEM_MARKER = "*";
+
+/**
+ * @internal
+ * Information about a path in the schema that requires coercion.
+ */
+export type CoercionTarget = {
+	path: string[];
+	type: "primitive" | "array" | "object" | "date";
+};
+
+/**
+ * Options for coercion behavior.
+ */
+export type CoerceOptions = {
+	/**
+	 * format to use for array parsing
+	 * @default "comma"
+	 */
+	arrayFormat?: "comma" | "json";
+};
+
+/**
+ * Recursively find all paths in a JSON Schema that require coercion.
+ * We prioritize "number", "integer", "boolean", "array", "object", and "date" types.
+ */
+export const findCoercionPaths = (
+	node: Record<string, any>,
+	path: string[] = [],
+): CoercionTarget[] => {
+	const results: CoercionTarget[] = [];
+
+	if (typeof node === "boolean") {
+		return results;
+	}
+
+	if ("const" in node) {
+		if (typeof node.const === "number" || typeof node.const === "boolean") {
+			results.push({ path: [...path], type: "primitive" });
+		}
+	}
+
+	if ("enum" in node && node.enum) {
+		if (
+			node.enum.some((v: any) => typeof v === "number" || typeof v === "boolean")
+		) {
+			results.push({ path: [...path], type: "primitive" });
+		}
+	}
+
+	if ("type" in node) {
+		if (node.type === "number" || node.type === "integer") {
+			results.push({ path: [...path], type: "primitive" });
+		} else if (node.type === "boolean") {
+			results.push({ path: [...path], type: "primitive" });
+		} else if (node.type === "string" && "format" in node && (node.format === "date-time" || node.format === "date")) {
+			results.push({ path: [...path], type: "date" });
+		} else if (node.type === "object") {
+			// Check if this object has properties defined
+			// If it does, we want to coerce the whole object from a JSON string
+			// But we also want to recursively check nested properties
+			const hasProperties =
+				"properties" in node &&
+				node.properties &&
+				Object.keys(node.properties).length > 0;
+
+			if (hasProperties) {
+				// Mark this path as needing object coercion (JSON parsing)
+				results.push({ path: [...path], type: "object" });
+			}
+
+			// Also recursively check nested properties for their own coercions
+			if ("properties" in node && node.properties) {
+				for (const [key, prop] of Object.entries(node.properties)) {
+					results.push(
+						...findCoercionPaths(prop as Record<string, any>, [...path, key]),
+					);
+				}
+			}
+		} else if (node.type === "array") {
+			// Mark the array itself as a target for splitting strings
+			results.push({ path: [...path], type: "array" });
+
+			if ("items" in node && node.items) {
+				if (Array.isArray(node.items)) {
+					// Tuple traversal
+					node.items.forEach((item, index) => {
+						results.push(
+							...findCoercionPaths(item as Record<string, any>, [...path, `${index}`]),
+						);
+					});
+				} else {
+					// List traversal
+					results.push(
+						...findCoercionPaths(node.items as Record<string, any>, [
+							...path,
+							ARRAY_ITEM_MARKER,
+						]),
+					);
+				}
+			}
+		}
+	}
+
+	if ("anyOf" in node && node.anyOf) {
+		for (const branch of node.anyOf) {
+			results.push(...findCoercionPaths(branch as Record<string, any>, path));
+		}
+	}
+
+	if ("allOf" in node && node.allOf) {
+		for (const branch of node.allOf) {
+			results.push(...findCoercionPaths(branch as Record<string, any>, path));
+		}
+	}
+
+	if ("oneOf" in node && node.oneOf) {
+		for (const branch of node.oneOf) {
+			results.push(...findCoercionPaths(branch as Record<string, any>, path));
+		}
+	}
+
+	// Deduplicate by path and type combination
+	const seen = new Set<string>();
+	return results.filter((t) => {
+		const key = JSON.stringify(t.path) + t.type;
+		if (seen.has(key)) return false;
+		seen.add(key);
+		return true;
+	});
+};
+
+/**
+ * Apply coercion to a data object based on identified paths.
+ */
+export const applyCoercion = (
+	data: unknown,
+	targets: CoercionTarget[],
+	options: CoerceOptions = {},
+) => {
+	const { arrayFormat = "comma" } = options;
+
+	// Helper to split string to array
+	const splitString = (val: string) => {
+		if (arrayFormat === "json") {
+			try {
+				return JSON.parse(val);
+			} catch {
+				return val;
+			}
+		}
+
+		if (!val.trim()) return [];
+		return val.split(",").map((s) => s.trim());
+	};
+
+	if (typeof data !== "object" || data === null) {
+		// If root data needs coercion
+		if (targets.some((t) => t.path.length === 0)) {
+			const rootTarget = targets.find((t) => t.path.length === 0);
+
+			if (rootTarget?.type === "object" && typeof data === "string") {
+				return coerceJson(data);
+			}
+
+			if (rootTarget?.type === "array" && typeof data === "string") {
+				return splitString(data);
+			}
+
+			if (rootTarget?.type === "date" && typeof data === "string") {
+				return coerceDate(data);
+			}
+
+			const asNumber = coerceNumber(data);
+			if (typeof asNumber === "number") {
+				return asNumber;
+			}
+			return coerceBoolean(data);
+		}
+		return data;
+	}
+
+	// Sort targets by path length to ensure parent objects/arrays are coerced before their children
+	const sortedTargets = [...targets].sort(
+		(a, b) => a.path.length - b.path.length,
+	);
+
+	const walk = (
+		current: unknown,
+		targetPath: string[],
+		type: "primitive" | "array" | "object" | "date",
+	) => {
+		if (!current || typeof current !== "object") return;
+
+		if (targetPath.length === 0) {
+			return;
+		}
+
+		// If we've reached the last key, apply coercion
+		if (targetPath.length === 1) {
+			const lastKey = targetPath[0];
+
+			if (lastKey === ARRAY_ITEM_MARKER) {
+				if (Array.isArray(current)) {
+					for (let i = 0; i < current.length; i++) {
+						const original = current[i];
+						if (type === "primitive") {
+							const asNumber = coerceNumber(original);
+							if (typeof asNumber === "number") {
+								current[i] = asNumber;
+							} else {
+								current[i] = coerceBoolean(original);
+							}
+						} else if (type === "object") {
+							current[i] = coerceJson(original);
+						} else if (type === "date") {
+							current[i] = coerceDate(original);
+						}
+					}
+				}
+				return;
+			}
+
+			const record = current as Record<string, unknown>;
+			// biome-ignore lint/suspicious/noPrototypeBuiltins: ES2020 compatibility
+			if (Object.prototype.hasOwnProperty.call(record, lastKey)) {
+				const original = record[lastKey];
+
+				if (type === "array" && typeof original === "string") {
+					record[lastKey] = splitString(original);
+					return;
+				}
+
+				if (type === "object" && typeof original === "string") {
+					record[lastKey] = coerceJson(original);
+					return;
+				}
+
+				if (type === "date" && typeof original === "string") {
+					record[lastKey] = coerceDate(original);
+					return;
+				}
+
+				if (Array.isArray(original)) {
+					if (type === "primitive") {
+						for (let i = 0; i < original.length; i++) {
+							const item = original[i];
+							const asNumber = coerceNumber(item);
+							if (typeof asNumber === "number") {
+								original[i] = asNumber;
+							} else {
+								original[i] = coerceBoolean(item);
+							}
+						}
+					}
+				} else {
+					if (type === "primitive") {
+						const asNumber = coerceNumber(original);
+						// If numeric parsing didn't produce a number, try boolean coercion
+						if (typeof asNumber === "number") {
+							record[lastKey] = asNumber;
+						} else {
+							record[lastKey] = coerceBoolean(original);
+						}
+					}
+				}
+			}
+			return;
+		}
+
+		// Recurse down
+		const [nextKey, ...rest] = targetPath;
+
+		if (nextKey === ARRAY_ITEM_MARKER) {
+			if (Array.isArray(current)) {
+				for (const item of current) {
+					walk(item, rest, type);
+				}
+			}
+			return;
+		}
+
+		const record = current as Record<string, unknown>;
+		walk(record[nextKey], rest, type);
+	};
+
+	for (const target of sortedTargets) {
+		walk(data, target.path, target.type);
+	}
+
+	return data;
+};
